# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Agent OS API
  version: 0.6.0
  description: >-
    Control plane and observability API for the Agent OS SPOG. Route
    surface is generated authoritatively from the chi router; payload
    shapes are defined by the SQL migrations (internal/db/migrations)
    and the sqlc query layer (internal/db/queries).
servers:
  - url: http://localhost:8080
    description: Local / hpms1 deployment
...
paths:
...
  /api/artifacts:
    get:
      tags: [artifacts]
      operationId: get_ListArtifacts
      summary: List Artifacts
      description: Handler ListArtifacts in internal/api.
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '404': { description: Not found }
        '500': { description: Server error }
    post:
      tags: [artifacts]
      operationId: post_UploadArtifact
      summary: Upload Artifact
      description: Handler UploadArtifact in internal/api.
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '404': { description: Not found }
        '500': { description: Server error }
  /api/artifacts/{id}:
    delete:
      tags: [artifacts]
      operationId: delete_DeleteArtifact
      summary: Delete Artifact
      description: Handler DeleteArtifact in internal/api.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '404': { description: Not found }
        '500': { description: Server error }
    get:
      tags: [artifacts]
      operationId: get_GetArtifact
      summary: Get Artifact
      description: Handler GetArtifact in internal/api.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '404': { description: Not found }
        '500': { description: Server error }
...
  /api/artifacts/{id}/file:
    get:
      tags: [artifacts]
      operationId: get_GetArtifactFile
      summary: Get Artifact File
      description: Handler GetArtifactFile in internal/api.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { description: Bad request }
        '404': { description: Not found }
        '500': { description: Server error }
...